+++ frontend/src/components/modals/guild-settings/guild-settings.tsx
@@ -18,19 +18,19 @@ const GuildSettings: React.FunctionComponent = () => {
     <Modal typeName={'GuildSettings'} size="full">
       <div className="grid grid-cols-12 h-full">
         <div className="col-span-4 bg-bg-secondary">
-          <nav className="float-right flex-grow py-14 w-48 my-1 mr-4">
+          <nav className="float-right flex-grow py-14 w-1/2 my-1 mr-4">
             <Category
               className="muted px-2.5 pb-1.5"
               title={guild.name} />
-          <NavTabs
-            tabs={[
-              { perm: 'MANAGE_GUILD', name: 'Overview', id: 'overview' },
-              { perm: 'MANAGE_ROLES', name: 'Roles', id: 'roles' },
-              { perm: 'MANAGE_INVITES', name: 'Invites', id: 'invites' },
-            ]}
-            tab={tab}
-            setTab={setTab}
-            predicate={t => perms.can(t.perm as any, guild.id)} />
+            <NavTabs
+              tabs={[
+                { perm: 'MANAGE_GUILD', name: 'Overview', id: 'overview' },
+                { perm: 'MANAGE_ROLES', name: 'Roles', id: 'roles' },
+                { perm: 'MANAGE_INVITES', name: 'Invites', id: 'invites' },
+              ]}
+              tab={tab}
+              setTab={setTab}
+              predicate={t => perms.can(t.perm as any, guild.id)} />
           </nav>
         </div>
 
@@ -47,5 +47,5 @@ const GuildSettings: React.FunctionComponent = () => {
     </Modal>
   ) : null;
 };
- 
+
 export default GuildSettings;