+++ frontend/src/components/modals/user-settings/user-settings.tsx
@@ -22,7 +22,7 @@ const UserSettings: React.FunctionComponent = () => {
       size="full">
       <div className="grid grid-cols-12 h-full">
         <div className="col-span-4 bg-bg-secondary">
-          <nav className="float-right flex-grow py-14 w-48 my-1 mr-4">
+          <nav className="float-right flex-grow py-14 w-1/2 my-1 mr-4">
             <Category
               className="normal px-2.5 pb-1.5"
               title="User Settings" />