+++ frontend/src/services/mention-service.ts
@@ -1,9 +1,7 @@
 import { getChannel, getChannelByName } from '../store/channels';
-import { getTag, getUser, getUserByTag } from '../store/users';
+import { getUser, getUserByTag, getUserByUsername } from '../store/users';
 
 export class MentionService {
-  public readonly tags = ['@245538070684827648', '#\d{18}'];
-
   private readonly patterns = {
     formatted: {
       channel: /<#(\d{18})>/gm,
@@ -12,21 +10,25 @@ export class MentionService {
     original: {
       channel: /#([A-Za-z\-\d]{2,32})/gm,
       user: /@([A-Za-z\d\-\_ ]{2,32}#\d{4})/gm,
+      userShorthand: /@([A-Za-z\d\-\_ ]{2,32})/gm,
     },
   };
 
   constructor(private state: Store.AppState) { }
 
-  // messageBox.onInput -> formatted mentions appear fancy in message box 
   public formatOriginal(content: string) {
     const guildId = this.state.ui.activeGuild!.id;
     return content
-      .replace(this.patterns.original.channel, (og, name) => {
+      .replaceAll(this.patterns.original.channel, (og, name) => {
         const channel = getChannelByName(guildId, name)(this.state);
-        return (channel) ? `<#${channel?.id}>` : og;
+        return (channel) ? `<#${channel.id}>` : og;
       })
-      .replace(this.patterns.original.user, (og, tag) => {
+      .replaceAll(this.patterns.original.user, (og, tag) => {
         const user = getUserByTag(tag)(this.state);
+        return (user) ? `<@${user.id}>` : og;
+      })
+      .replaceAll(this.patterns.original.userShorthand, (og, tag) => {
+        const user = getUserByUsername(tag)(this.state);
         return (user) ? `<@${user.id}>` : og;
       });
   }
@@ -39,22 +41,31 @@ export class MentionService {
 
   private mentionAnchorTag(type: 'channel' | 'user', id: string) {
     const selfUserId = this.state.auth.user!.id;
-    const guildId = this.state.ui.activeGuild!.id;
+    const channelId = this.state.ui.activeChannel?.id;
+    const guildId = this.state.ui.activeGuild?.id;
+
+    var channel = getChannel(id)(this.state);
+    var user = getUser(id)(this.state);
+
     const tag = {
       channel: {
-        onClick: `window.location.href = '/channels/${guildId}/${id}'`,
-        text: `#${getChannel(id)(this.state)?.name}`,
+        onClick: `window.location.href = '/channels/${guildId}/${id}';`,
+        text: `#${channel?.name}`,
       },
       user: {
         onClick: `events.emit('openUserProfile', '${id}')`,
-        text: `@${getTag(getUser(id)(this.state))}`,
+        text: `@${user?.username}`,
       },
     };
 
-    const mentioned = (id === selfUserId) ? 'bg-tertiary rounded px-1' : '';
-    return `<a
-      data-id="${id}"
-      class="font-extrabold cursor-pointer hover:underline ${mentioned}"
-      onclick="${tag[type].onClick}">${tag[type].text}</a>`;
+    const mentioned = (id === selfUserId || id === channelId)
+      ? 'bg-tertiary rounded px-1'
+      : 'bg-bg-tertiary rounded px-1';
+
+    return (user || channel)
+      ? `<a data-id="${id}"
+        class="font-extrabold cursor-pointer hover:underline ${mentioned}"
+        onclick="${tag[type].onClick}">${tag[type].text}</a>`
+      : `<a>Not Found</a>`;
   }
 }